+++ .eslintrc.js
@@ -6,13 +6,19 @@ module.exports = {
     mocha: true,
     node: true,
   },
-  extends: ['eslint:recommended', 'google'],
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     ecmaVersion: 12,
   },
+  extends: ['eslint:recommended', 'google', 'plugin:@typescript-eslint/eslint-recommended', 'plugin:@typescript-eslint/recommended'],
   rules: {
-    'arrow-parens': ['error', 'as-needed'],
-    'max-len': ['error', { code: 160, ignoreUrls: true }],
+    'prettier/prettier': 'error',
+    'arrow-parens': ['error', 'always'],
+    'max-len': ['error', { code: 140, ignoreUrls: true }],
     'object-curly-spacing': ['off'],
+    // eslint-disable-next-line prettier/prettier
+    'indent': ['error', 2, { SwitchCase: 1 }],
   },
+  ignorePatterns: ['lib'],
+  plugins: ['prettier'],
 };