+++ src/types/table.ts
@@ -1,0 +1,2 @@
+export type SecurityReportHeader = 'ID' | 'Module' | 'Title' | 'Sev.' | 'URL' | 'Ex.';
+export type ExceptionReportHeader = 'ID' | 'Status' | 'Expiry' | 'Notes';