+++ src/utils/print.ts
@@ -1,15 +1,16 @@
-const table = require('table').table;
+import { table, TableUserConfig } from 'table';
+import { SecurityReportHeader, ExceptionReportHeader } from 'src/types';
 
-const SECURITY_REPORT_HEADER = ['ID', 'Module', 'Title', 'Sev.', 'URL', 'Ex.'];
-const EXCEPTION_REPORT_HEADER = ['ID', 'Status', 'Expiry', 'Notes'];
+const SECURITY_REPORT_HEADER: SecurityReportHeader[] = ['ID', 'Module', 'Title', 'Sev.', 'URL', 'Ex.'];
+const EXCEPTION_REPORT_HEADER: ExceptionReportHeader[] = ['ID', 'Status', 'Expiry', 'Notes'];
 
 /**
  * Print the security report in a table format
  * @param  {Array} data   Array of arrays
  * @return {undefined}    Returns void
  */
-function printSecurityReport(data) {
-  const configs = {
+export function printSecurityReport(data: string[][]): void {
+  const configs: TableUserConfig = {
     singleLine: true,
     header: {
       alignment: 'center',
@@ -25,8 +26,8 @@ function printSecurityReport(data) {
  * @param  {Array} data   Array of arrays
  * @return {undefined}    Returns void
  */
-function printExceptionReport(data) {
-  const configs = {
+export function printExceptionReport(data: string[][]): void {
+  const configs: TableUserConfig = {
     singleLine: true,
     header: {
       alignment: 'center',
@@ -36,8 +37,3 @@ function printExceptionReport(data) {
 
   console.info(table([EXCEPTION_REPORT_HEADER, ...data], configs));
 }
-
-module.exports = {
-  printSecurityReport,
-  printExceptionReport,
-};