+++ test/utils/common.test.ts
@@ -1,7 +1,5 @@
-const chai = require('chai');
-const { expect } = chai;
-
-const { isWholeNumber, isJsonString } = require('../../utils/common');
+import { expect } from 'chai';
+import { isWholeNumber, isJsonString } from '../../src/utils/common';
 
 describe('Common utils', () => {
   describe('#isJsonString', () => {
@@ -16,7 +14,7 @@ describe('Common utils', () => {
 
   describe('#isWholeNumber', () => {
     it('should be able to determine a whole number', () => {
-      expect(isWholeNumber()).to.equal(false);
+      expect(isWholeNumber(undefined)).to.equal(false);
       expect(isWholeNumber(0.14)).to.equal(false);
       expect(isWholeNumber(20.45)).to.equal(false);
       expect(isWholeNumber('')).to.equal(false);
@@ -28,8 +26,8 @@ describe('Common utils', () => {
       expect(isWholeNumber(2920)).to.equal(true);
       expect(isWholeNumber(934)).to.equal(true);
       expect(isWholeNumber('0920')).to.equal(true);
-
-      expect(isWholeNumber(true)).to.equal(true); // Should handle this?
     });
   });
 });
+
+export {};